requirements.txt
mautrix
aiohttp
PyYAML
pytz
asyncpg
SQLAlchemy
alembic
markdown
Pillow
python-magic
commonmark
ruamel.yaml
